+++ posthog-ai/src/vercel/middleware.ts
@@ -129,7 +129,7 @@ const mapVercelPrompt = (prompt: LanguageModelV1Prompt): PostHogInput[] => {
     let serialized = JSON.stringify(inputs)
     let removedCount = 0
     // We need to keep track of the initial size of the inputs array because we're going to be mutating it
-    let initialSize = inputs.length
+    const initialSize = inputs.length
     for (let i = 0; i < initialSize && Buffer.byteLength(serialized, 'utf8') > MAX_OUTPUT_SIZE; i++) {
       inputs.shift()
       removedCount++