+++ posthog-node/src/extensions/error-tracking/types.ts
@@ -1,6 +1,8 @@
 // Portions of this file are derived from getsentry/sentry-javascript by Software, Inc. dba Sentry
 // Licensed under the MIT License
 
+import { JsonType } from 'posthog-core/src/types'
+
 // levels originally copied from Sentry to work with the sentry integration
 // and to avoid relying on a frequently changing @sentry/types dependency
 // but provided as an array of literal types, so we can constrain the level below
@@ -64,6 +66,6 @@ export interface StackFrame {
   in_app?: boolean
   instruction_addr?: string
   addr_mode?: string
-  vars?: { [key: string]: any }
+  vars?: { [key: string]: JsonType }
   debug_id?: string
 }