+++ posthog-react-native/src/posthog-rn.ts
@@ -11,6 +11,7 @@ import {
   PostHogPersistedProperty,
   SurveyResponse,
   logFlushError,
+  maybeAdd,
 } from '../../posthog-core/src'
 import { getLegacyValues } from './legacy'
 import { PostHogRNStorage, PostHogRNSyncMemoryStorage } from './storage'
@@ -195,7 +196,7 @@ export class PostHog extends PostHogCore {
     return `${this.getLibraryId()}/${this.getLibraryVersion()}`
   }
 
-  getCommonEventProperties(): any {
+  getCommonEventProperties(): PostHogEventProperties {
     return {
       ...super.getCommonEventProperties(),
       ...this._appProperties,
@@ -205,7 +206,7 @@ export class PostHog extends PostHogCore {
   }
 
   // Custom methods
-  async screen(name: string, properties?: { [key: string]: any }, options?: PostHogCaptureOptions): Promise<void> {
+  async screen(name: string, properties?: PostHogEventProperties, options?: PostHogCaptureOptions): Promise<void> {
     await this._initPromise
     // Screen name is good to know for all other subsequent events
     this.registerForSession({
@@ -459,11 +460,13 @@ export class PostHog extends PostHogCore {
 
     // version and build are deprecated, but we keep them for compatibility
     // use $app_version and $app_build instead
-    const properties: PostHogEventProperties = { version: appVersion, build: appBuild }
+    const properties: PostHogEventProperties = { ...maybeAdd('version', appVersion), ...maybeAdd('build', appBuild) }
 
     if (!isMemoryPersistence) {
-      const prevAppBuild = this.getPersistedProperty(PostHogPersistedProperty.InstalledAppBuild)
-      const prevAppVersion = this.getPersistedProperty(PostHogPersistedProperty.InstalledAppVersion)
+      const prevAppBuild = this.getPersistedProperty(PostHogPersistedProperty.InstalledAppBuild) as string | undefined
+      const prevAppVersion = this.getPersistedProperty(PostHogPersistedProperty.InstalledAppVersion) as
+        | string
+        | undefined
 
       if (!appBuild || !appVersion) {
         this.logMsgIfDebug(() =>
@@ -481,8 +484,8 @@ export class PostHog extends PostHogCore {
         } else if (prevAppBuild !== appBuild) {
           // app updated
           this.capture('Application Updated', {
-            previous_version: prevAppVersion,
-            previous_build: prevAppBuild,
+            ...maybeAdd('previous_version', prevAppVersion),
+            ...maybeAdd('previous_build', prevAppBuild),
             ...properties,
           })
         }
@@ -499,7 +502,7 @@ export class PostHog extends PostHogCore {
 
     this.capture('Application Opened', {
       ...properties,
-      url: initialUrl,
+      ...maybeAdd('url', initialUrl),
     })
 
     AppState.addEventListener('change', (state) => {