+++ posthog-react-native/test/posthog.spec.ts
@@ -80,7 +80,6 @@ describe('PostHog React Native', () => {
     // await posthog.ready()
 
     expect(posthog.getCommonEventProperties()).toEqual({
-      $active_feature_flags: undefined,
       $lib: 'posthog-react-native',
       $lib_version: expect.any(String),
       $screen_height: expect.any(Number),
@@ -100,7 +99,6 @@ describe('PostHog React Native', () => {
     await posthog.ready()
 
     expect(posthog2.getCommonEventProperties()).toEqual({
-      $active_feature_flags: undefined,
       $lib: 'posthog-react-native',
       $lib_version: expect.any(String),
       $screen_height: expect.any(Number),